fix: bug fixes

+++ utils/gitManagers/plugins/github/auth/handleDisconnect.js
@@ -1,8 +1,12 @@
 const { configstore } = require('../../../../../configstore')
 
-const disconnect = async () => {
+const handleDisconnect = async (options, config) => {
+  if (!config.userId && !config.userToken) {
+    throw new Error('No connected user found for github')
+  }
+
   configstore.delete('githubUserId')
   configstore.delete('githubUserToken')
 }
 
-module.exports = disconnect
+module.exports = handleDisconnect
